lambda/pyproject.toml
[tool.poetry]
name = "kbot"
version = "0.1.0"
description = ""
authors = ["mozkzki <mozkzki@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "2.25.1"
line-bot-sdk = "1.16.0"
boto3 = "1.14.23"

[tool.poetry.dev-dependencies]
flake8 = "4.0.1"
mypy = "0.910"
pytest = "6.2.5"
black = "21.9b0"
pytest-cov = "3.0.0"
coverage = "6.0.2"
isort = "5.9.3"
regex = "2021.9.30"
types-requests = "^2.25.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
